# api/swagger.json
{
  "basePath": "/api/2.0",
  "consumes": [
    "application/json"
  ],
  "definitions": {
    "Error": {
      "properties": {
        "code": {
          "format": "int32",
          "type": "integer"
        },
        "fields": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Action": {
      "description": "Action object",
      "properties": {
        "command": {
          "description": "Command to execute",
          "enum": [
            "cancel"
          ],
          "type": "string"
        },
        "options": {
          "description": "Command options object",
          "type": "object"
        }
      },
      "required": [
        "command"
      ]
    },
...
    "PostWorkflow": {
      "properties": {
        "name": {
          "type": "string"
        },
        "options": {
          "properties": {
            "default": {
              "properties": {
                "graphOptions": {
                  "type": "object"
                },
                "nodeId": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "type": "object"
        }
      },
      "type": "object"
    },
...
    "notification_obj": {
      "properties": {
        "nodeId": {
          "type": "string"
        }
      }
    }
  },
  "info": {
    "title": "onTaskClient",
    "version": "0.0.1"
  },
  "paths": {
...
    "/workflows": {
      "get": {
        "description": "Get list workflow that have been run, or are currently running.\n",
        "operationId": "workflowsGet",
        "parameters": [{
          "description": "Query string to search for specific workflow properties\n",
          "in": "query",
          "name": "active",
          "required": false,
          "type": "boolean"
        },
          {
            "description": "Query string specifying properties to search for",
            "in": "query",
            "name": "$skip",
            "required": false,
            "type": "integer"
          },
          {
            "description": "Query string specifying properties to search for",
            "in": "query",
            "name": "$top",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull retrieved workflows\n",
            "schema": {
              "type": "object"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "summary": "Get a list of workflow instances\n",
        "tags": [
          "workflows"
        ],
        "x-authentication-type": [
          "jwt"
        ],
        "x-privileges": [
          "Read",
          "workflowsRead"
        ],
        "x-view": "workflows.2.0.json"
      },
      "post": {
        "description": "Run a workflow by specifying a workflow graph injectable name.\nThe workflow is not associated with a node.\n",
        "operationId": "workflowsPost",
        "parameters": [{
          "description": "The body must contain the name property with the\nvalue set to the injectable name of the workflow\ngraph to run\n",
          "in": "body",
          "name": "body",
          "required": false,
          "schema": {
            "$ref": "#/definitions/PostWorkflow"
          }
        },
          {
            "description": "pass node id",
            "in": "query",
            "name": "nodeId",
            "required": false,
            "type": "string"
          },

          {
            "description": "Optionally specify the injectable name of the workflow graph to run",
            "in": "query",
            "name": "name",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully posted the workflow\n",
            "schema": {
              "type": "object"
            }
          },
          "500": {
            "description": "Workflow was not run\n",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Upload failed",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "summary": "Run a workflow\n",
        "tags": [
          "workflows"
        ],
        "x-authentication-type": [
          "jwt"
        ],
        "x-privileges": [
          "Write",
          "workflowsWrite"
        ]
      },
      "x-swagger-router-controller": "workflows"
    },
...
    "/workflows/{identifier}": {
      "delete": {
        "description": "Delete the workflow with the specified instance identifier.\n",
        "operationId": "workflowsDeleteByInstanceId",
        "parameters": [{
          "description": "The workflow instance identifier\n",
          "in": "path",
          "name": "identifier",
          "required": true,
          "type": "string"
        }],
        "responses": {
          "204": {
            "description": "Successfully deleted the specified workflow\n",
            "schema": {
              "type": "object"
            }
          },
          "404": {
            "description": "The specified workflow was not found\n",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "summary": "Delete a workflow\n",
        "tags": [
          "workflows"
        ],
        "x-authentication-type": [
          "jwt"
        ],
        "x-privileges": [
          "Write",
          "workflowsRemove"
        ]
      },
      "get": {
        "description": "Get the workflow with the specified instance identifier.\n",
        "operationId": "workflowsGetByInstanceId",
        "parameters": [{
          "description": "The workflow instance identifier\n",
          "in": "path",
          "name": "identifier",
          "required": true,
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "Successfully retrieved the specified workflow\n",
            "schema": {
              "type": "object"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "summary": "Get a workflow\n",
        "tags": [
          "workflows"
        ],
        "x-authentication-type": [
          "jwt"
        ],
        "x-privileges": [
          "Read",
          "workflowsRead"
        ],
        "x-view": "workflows.2.0.json"
      },
      "x-swagger-router-controller": "workflows"
    },
    "/workflows/{identifier}/action": {
      "put": {
        "description": "Perform the specified action on the workflow with the specified\ninstance identifier. Currently, the cancel action is supported.\n",
        "operationId": "workflowsAction",
        "parameters": [{
          "description": "The workflow instance identifier\n",
          "in": "path",
          "name": "identifier",
          "required": true,
          "type": "string"
        },
          {
            "description": "The action property with the value of the desired action",
            "in": "body",
            "name": "action",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Action"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successfully performed the action on the specified workflow\n",
            "schema": {
              "type": "object"
            }
          },
          "404": {
            "description": "The workflow with the identifier was not found\n",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "summary": "Perform an action on a workflow\n",
        "tags": [
          "workflows"
        ],
        "x-authentication-type": [
          "jwt"
        ],
        "x-privileges": [
          "Write",
          "workflowsModify"
        ]
      },
      "x-swagger-router-controller": "workflows"
    },
...
    "/notification": {
      "x-swagger-router-controller": "notification",
      "post": {
        "operationId": "notificationPost",
        "summary": "post a notification\n",
        "description": "post a notification\n",
        "parameters": [{
          "name": "nodeId",
          "in": "query",
          "description": "Node instance identifier\n",
          "required": false,
          "type": "string"
        },
          {
            "name": "data",
            "in": "body",
            "description": "notification data\n",
            "required": false,
            "schema": {
              "$ref": "#/definitions/notification_obj"
            }
          }
        ],
        "tags": [
          "southBound"
        ],
        "responses": {
          "201": {
            "description": "Specifics of the notification\n",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "bad request parameter passed or no active task or taskgraph.\n",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/notification/progress": {
      "x-swagger-router-controller": "notification",
      "get": {
        "operationId": "notificationProgressPost",
        "summary": "Send progress notification to task\n",
        "description": "This API is used to update the progress information for a task instance\n",
        "parameters": [{
          "name": "taskId",
          "in": "query",
          "description": "The identifier of task instance which the progress applies to\n",
          "required": false,
          "type": "string"
        },
          {
            "name": "maximum",
            "in": "query",
            "description": "The maximum progress value\n",
            "required": false,
            "type": "string"
          },
          {
            "name": "value",
            "in": "query",
            "description": "The current progress value\n",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "southBound"
        ],
        "responses": {
          "201": {
            "description": "Specifics of the notification\n",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "bad request parameter passed or no active task or taskgraph.\n",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "operationId": "notificationProgressPost",
        "summary": "Send progress notification to task\n",
        "description": "This API is used to update the progress information for a task instance\n",
        "parameters": [{
          "name": "taskId",
          "in": "query",
          "description": "The identifier of task instance which the progress applies to\n",
          "required": false,
          "type": "string"
        },
          {
            "name": "maximum",
            "in": "query",
            "description": "The maximum progress value\n",
            "required": false,
            "type": "string"
          },
          {
            "name": "value",
            "in": "query",
            "description": "The current progress value\n",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "southBound"
        ],
        "responses": {
          "201": {
            "description": "Specifics of the notification\n",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "bad request parameter passed or no active task or taskgraph.\n",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
...
}